Organizational changes, including separating out consts and enums; keeping main just for imports and main.
