+++ Cargo.toml
@@ -8,15 +8,15 @@ edition = "2018"
 name="rustBCA"
 
 [dependencies]
-rand = "0.7.3"
-geo = "0.14.2"
-toml = "0.5.6"
-anyhow = "1.0.32"
-itertools = "0.9.0"
-rayon = "1.4.0"
-indicatif = {version="*", features=["rayon"]}
-serde = { version = "1.0.116", features = ["derive"] }
-hdf5 = {version = "0.7.0", optional = true}
+rand = "0.8.3"
+geo = "0.17.1"
+toml = "0.5.8"
+anyhow = "1.0.38"
+itertools = "0.10.0"
+rayon = "1.5.0"
+indicatif = {version="0.15.0", features=["rayon"]}
+serde = { version = "1.0.123", features = ["derive"] }
+hdf5 = {version = "0.7.1", optional = true}
 openblas-src = {version  = "0.9", optional = true}
 netlib-src = {version  = "0.8", optional = true}
 intel-mkl-src = {version = "0.6.0", optional = true}
@@ -30,7 +30,7 @@ float-cmp = "0.8.0"
 lto = "fat"
 codegen-units = 1
 opt-level = 3
-debug = true
+debug = false
 
 [features]
 hdf5_input = ["hdf5"]
